+++ common/cpp/src/database/mongodb_client.h
@@ -58,10 +58,12 @@ class MongoDBClient final : public Database {
     Error GetLastStream(StreamInfo* info) const override;
     Error DeleteStream(const std::string& stream) const override;
     Error GetMetaFromDb(const std::string& collection, const std::string& id, std::string* res) const override;
+    Error GetNextId(const std::string& collection, uint64_t* id) const;
     ~MongoDBClient() override;
   private:
     mongoc_client_t* client_{nullptr};
     mutable mongoc_collection_t* current_collection_{nullptr};
+    mutable mongoc_client_session_t* current_session_{nullptr};
     mutable std::string current_collection_name_;
     std::string database_name_;
     mongoc_write_concern_t* write_concern_{nullptr};
@@ -82,6 +84,15 @@ class MongoDBClient final : public Database {
     Error DeleteCollection(const std::string& name) const;
     Error DeleteCollections(const std::string& prefix) const;
     Error DeleteDocumentsInCollection(const std::string& collection_name, const std::string& querystr) const;
+    Error InsertWithAutoId(const MessageMeta& file, const std::string& collection, bool ignore_duplicates) const;
+};
+
+struct TransactionContext {
+    const MongoDBClient* caller;
+    MessageMeta meta;
+    std::string collection;
+    bool ignore_duplicates;
+    Error err;
 };
 
 }